feat: Implement authorization for blog posts
This change adds authorization to blog posts so that only the author can edit or delete their own posts.

A new middleware, `CheckPostOwner`, has been created to check if the authenticated user is the author of the post. If the user is not the author, they will be redirected to a 403 error page.

The `BlogPostController` has been updated to use the new middleware on the `edit` and `update` methods.